following last submission, fix the __asm parsing.
now it can parse __asm{} or single __asm statement. still it can't parse mixed __asm statement for example,

__asm mov blah;
int a;

in this case the second declaration will be omitted. but it doesn't look like a big problem at this stage.

also the code is moved from compound_statement to expression_statement to unify asm parsing.