first working version to batch compile files
I find out that for batching compile to work the stages must be isolated - that is the lexer, parser, symbol table state must be reset every time a source file is compiled.

previously it doesn't work well because symbol table state is not properly reset. this submission fixed this.

also this submission fixed the logging system when entering new context (entering or leaving an included file in preprocessing stage).