refine enum type checking.
when define an enum, check if it's previously declared or defined. If it's previously declared but not defined, and it's declared as enum, then it's ok and a definition goes here. If it's previously declared as struct, then flag an error. On the other hand, if it's previously defined, then unless they are at different scope (plus previous is not an struct declare) otherwise flag an error.