Fix the parsing of external declaration.
A stand alone global external declaration, like this:
x;
would be a valid declarator with int type assumed.

As a result, declaration specifier is allowed to be null in semantic check phase to represent the "default assumed int" case, which I don't feel very well and might need to revisit and devise a more consistent representation later.