Add a special utility add_symbol_ssc for semantic analysis process.
In semantic analysis phase, it's not allowed to alter the symbol table level, this is different from the symbol table routine the parser using which do need to alter symbol level because of existing of enterscope and exitscop pairs.

To do - there is a bug in ssc enumerator checking where constant which is supposed to get folded doesn't. need to check!!